requirements.txt
streamlit==1.40.0
supabase==2.9.1
python-dotenv==1.0.1
reportlab==4.2.5
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
